Soluzione sporca ma obbligata per cambio dinamico immagini varianti prodotto

+++ sql/tdw.sql
@@ -119,11 +119,11 @@ CREATE TABLE product_image (
 CREATE TABLE offer (
     offer_id INTEGER UNSIGNED AUTO_INCREMENT PRIMARY KEY,
     `description` TEXT NULL,
-    percentage DECIMAL(5, 2) NOT NULL,
+    percentage SMALLINT NOT NULL,
     start_date DATETIME NOT NULL,
     end_date DATETIME NOT NULL,
     product_id INTEGER UNSIGNED NOT NULL,
-    CHECK (percentage BETWEEN 0.00 AND 100.00),
+    CHECK (percentage BETWEEN 0 AND 100),
     FOREIGN KEY (product_id)
         REFERENCES product (product_id)
         ON DELETE CASCADE ON UPDATE CASCADE,
